Added technology merging and buttons

+++ Assets/Scripts/Technology.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Technology", menuName = "Scriptable Objects/Technology")]
 public class Technology : ScriptableObject
 {
-    public Dictionary<int, float> stats = new Dictionary<int, float>(); //(0, health), (1,dmg), (2, abilityDamage), (3, atkspd), (4, castspd), (5, abilityCount), (6, luck)
+    public Dictionary<int, double> stats = new Dictionary<int, double>(); //(0, health), (1,dmg), (2, abilityDamage), (3, atkspd), (4, castspd), (5, abilityCount), (6, luck)
     public Dictionary<int, Ability> abilities = new Dictionary<int, Ability>();
     public Dictionary<int, Passive> passives = new Dictionary<int, Passive>();
 
-    public void Initialize(Dictionary<int, float> statDictionary,
+    public void Initialize(Dictionary<int, double> statDictionary,
                        Dictionary<int, Ability> abilityDictionary,
                        Dictionary<int, Passive> passiveDictionary)
     {
